# docs/swagger.yaml
definitions:
  handler.AuthResponse:
    properties:
      token:
        example: eyJhbGci...
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  handler.CreateUserRequest:
    properties:
      email:
        example: dev@company.com
        type: string
      password:
        example: supersecret123
        type: string
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: invalid JSON
        type: string
    type: object
  handler.IssueDetailResponse:
    properties:
      events:
        items:
          $ref: '#/definitions/handler.IssueEventRow'
        type: array
      issue:
        $ref: '#/definitions/models.Issue'
    type: object
  handler.IssueEventRow:
    properties:
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      payload: {}
    type: object
  handler.IssueListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.Issue'
        type: array
      limit:
        example: 50
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 143
        type: integer
    type: object
  handler.LoginRequest:
    properties:
      email:
        example: dev@company.com
        type: string
      password:
        example: supersecret
        type: string
    type: object
...
  handler.UpdateStatusRequest:
    properties:
      status:
        enum:
        - open
        - resolved
        - ignored
        example: resolved
        type: string
    type: object
...
  models.Issue:
    properties:
      error_message:
        type: string
      error_type:
        type: string
      fingerprint:
        type: string
      first_seen:
        type: string
      id:
        type: string
      last_seen:
        type: string
      level:
        type: string
      occurrences:
        type: integer
      project_id:
        type: string
      status:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/admin/users:
    post:
      consumes:
      - application/json
      description: Manually provision a new user account. Access is granted by the
        admin, not via self-registration. Requires X-Admin-Key header.
      parameters:
      - description: Admin key from ADMIN_KEY env var
        in: header
        name: X-Admin-Key
        required: true
        type: string
      - description: New user credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: invalid or missing admin key
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: email already registered
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create user
      tags:
      - admin
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password. Returns a JWT valid for 24
        hours.
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: invalid email or password
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Login
      tags:
      - auth
...
  /api/projects/{projectId}/issues:
    get:
      description: Returns a paginated list of deduplicated issues for a project,
        ordered by last_seen descending. Each item includes the occurrence counter
        so the frontend can show "occurred N times".
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: Max results (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Pagination offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.IssueListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List issues
      tags:
      - issues
  /api/projects/{projectId}/issues/{issueId}:
    get:
      description: Returns issue details and the last 10 raw event payloads that caused
        it.
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: Issue UUID
        in: path
        name: issueId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.IssueDetailResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get issue
      tags:
      - issues
    patch:
      consumes:
      - application/json
      description: Changes the status of an issue. Use "resolved" to mark it fixed,
        "ignored" to suppress it, or "open" to reopen.
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: Issue UUID
        in: path
        name: issueId
        required: true
        type: string
      - description: New status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Issue'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update issue status
      tags:
      - issues
...
swagger: "2.0"