# internal/httpServer/handler/helpers.go
package handler

import (
	"encoding/json"
	"net/http"
)

// ErrorResponse is the standard error body returned on all 4xx/5xx responses.
type ErrorResponse struct {
	Error string `json:"error" example:"invalid JSON"`
}

func writeError(w http.ResponseWriter, status int, message string) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	json.NewEncoder(w).Encode(ErrorResponse{Error: message})
}

func writeJSON(w http.ResponseWriter, status int, v any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	json.NewEncoder(w).Encode(v)
}

func decodeJSON(r *http.Request, v any) error {
	return json.NewDecoder(r.Body).Decode(v)
}